backend/requirements.txt
django==3.2.15
psycopg2-binary==2.9.3
djangorestframework==3.13.1
markdown
django-filter==22.1
djoser==2.1.0
Pillow==9.2.0
drf_extra_fields==3.4.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0
django-colorfield==0.7.2